[render] add MeshSelectorEditorSettings struct
